fix: テスト環境でのモジュール解決エラーを修正
- select変換関数をselect-helpers.tsに切り出し
- テストファイルをbun:testに変更（vitestから）
- @tanstack/react-start依存を避けてテスト可能に
