+++ apps/web/src/lib/transformers/chart-data.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from "vitest";
+import { describe, expect, it } from "bun:test";
 import type { SongStat, StackedWorkStat, WorkStat } from "@/lib/public-api";
 import {
 	transformSimpleDataForNivo,
